LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgnumber-animation-invalid-value-1-expected.txt


FAIL Test SVGNumber animation with invalid values: String values in from and to attributes. assert_equals: expected 1 but got 0
